@@ -5,15 +5,14 @@ apply plugin: 'kotlin-android-extensions'
 
  
 android {
-    compileSdkVersion 28
-
+    compileSdkVersion 34
     defaultConfig {
         applicationId "com.indiewalk.watchdog.earthquake"
         minSdkVersion 16
-        targetSdkVersion 28
+        targetSdkVersion 34
         versionCode 10
         versionName "2.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
         vectorDrawables.useSupportLibrary = true
     }
     buildTypes {
@@ -24,8 +23,24 @@ android {
     }
 
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_11
+        targetCompatibility JavaVersion.VERSION_11
+    }
+
+    kotlinOptions {
+        jvmTarget = JavaVersion.VERSION_11
+    }
+
+    lintOptions {
+        abortOnError false
+    }
+
+    dataBinding {
+        enabled true
+    }
+
+    viewBinding {
+        enabled true
     }
 
     testOptions {
