@@ -1,10 +1,10 @@
 package com.indiewalk.watchdog.earthquake.data.local
 
-import android.arch.lifecycle.LiveData
-import android.arch.persistence.room.Dao
-import android.arch.persistence.room.Insert
-import android.arch.persistence.room.OnConflictStrategy
-import android.arch.persistence.room.Query
+import androidx.lifecycle.LiveData
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
 import com.indiewalk.watchdog.earthquake.data.model.Earthquake
 
 @Dao
