@@ -1,7 +1,7 @@
 package com.indiewalk.watchdog.earthquake.data.remote
 
-import android.arch.lifecycle.LiveData
-import android.arch.lifecycle.MutableLiveData
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
 import android.content.Context
 import android.content.Intent
 import android.content.SharedPreferences
