@@ -1,7 +1,7 @@
 package com.indiewalk.watchdog.earthquake.ui
 
-import android.arch.lifecycle.LiveData
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.ViewModel
 import android.content.Context
 import android.content.SharedPreferences
 import android.preference.PreferenceManager
