@@ -1,7 +1,7 @@
 package com.indiewalk.watchdog.earthquake.ui
 
-import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 
 class MainViewModelFactory(private val listType: String) : ViewModelProvider.NewInstanceFactory() {
 
